requirements.txt
absl-py==0.10.0
anyio==3.3.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2019.11.28
charset-normalizer==2.0.12
click==7.1.1
colorama==0.4.4
crcmod==1.7
cycler==0.11.0
decorator==4.4.2
decord==0.6.0
defusedxml==0.6.0
dill==0.3.4
dm-tree==0.1.6
dnspython==1.16.0
entrypoints==0.3
eventlet==0.25.1
flatbuffers==1.12
fonttools==4.29.1
future==0.18.2
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.55.0
greenlet==0.4.15
grpcio==1.32.0
h5py==2.10.0
idna==3.3
importlib-metadata==1.5.0
importlib-resources==5.4.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-server==1.10.2
jupyterlab==3.1.4
jupyterlab-server==2.6.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.5.1
mistune==0.8.4
monotonic==1.5
nbclassic==0.3.1
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.2.0
object-detection==0.1
opencv-python==4.2.0.32
opt-einsum==3.3.0
orjson==3.6.4
packaging==21.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.4
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
Pygments==2.6.1
pyparsing==3.0.7
pyrsistent==0.15.7
python-dateutil==2.8.1
pywin32==303
pywinpty==2.0.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
sacrebleu==2.0.0
scikit-learn==0.22.2.post1
scipy==1.1.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-addons==0.12.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.4.0
tensorflow-hub==0.11.0
tensorflow-metadata==0.27.0
tensorflow-model-optimization==0.5.0
tensorflow-text==2.7.0rc1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.63.0
traitlets==4.3.3
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
